activate/requirements.txt
Django==3.0.3
django-crispy-forms==1.9.0
Pillow==7.0.0
pylint
django-filter